fix: close devices on signal
Currently, we are opening devices on program startup
when registering TorchDeviceImpl. tt-metal doesn't like
when the device is left open. So if pybuda crashes,
next start up of pybuda will be stuck on trying to open
devices - to mitigate this, we need to reset the board(s)
manually.

This change ensures that we close devices on crash.