[training] bootstrapping training (#99)
Initial changes for bootstrapping training.

- expanding `pybuda.compile` to accept loss and optimizer module.
- running autograd
- forward and backward graphs are lowered into tt-mlir as two
  separate functions - resulting binary has two programs (fwd and bwd)
- expanding `CompiledModel` with `backward()`

For initial training support we are going with executing loss and
optimizer modules on CPU.

Currently, the most problematic part is isolating and lowering
fwd and bwd parts of the graph, since the autograd is creating
a single graph. For now, I've hacked it up using
`GraphTraversalContext`. In changes to follow, we will separate
these two into different graphs. That should make things much
cleaner - e.g., each graph will have precisely defined inputs
and outputs unlike now, where we need to deduce what those
are for a particular subgraph.

Handling of intermediate results and storing them as outputs
of the forward pass to be used later in the backward pass
is NOT implemented. This will follow after the autograd
changes are done and we have two separate graphs.

For separating backward and forward executions, I've
experimented with two approaches:
1. creating two binaries
2. creating one binary with two programs

The second option seems cleaner and will simplify
handling of the binary, but in future we might support both approaches.

Basic test for training single parametrized multiply op is added.

Issues #17 #18 #19 #20