[Llama 3B] CPU samples of running model without using .generative() API. #185
Main purpose of this change is to demonstrate how how to perform inference
on the Llama 3B model on CPU without using the `.generate` function from
the transformers library.

Also, have in mind that tests are marked to be skipped in CI as they are
proof-of-concept (PoC) and should be mapped to work on the actual device.

Note: The actual implementation when running end-to-end (E2E) on the device
might differ depending on various constraints.