[ci] cleanup submodules when running CI (#193)
We've hit build issues when bumping tt-mlir version and
to resolve those #190 was merged. This is a follow up change
with a cleaner fix.

Github runner will reuse the same folder for repo checkout
between jobs. The checkout action from github cleans leftover
files, but not for git submodules within a repo. This creates
a problem, since it can happen that cmake doesn't regenerate
build files, but incorrectly, reuse ones that were left over.

This change enforces cleanup of git submodules, resulting in
correct, but slower builds.

Also, the build command is rewritten to be more explicit.