[fix] remove register device guard to avoid hang (#219)
On latest version of tt-metal, there is an added synchronization when
closing the device. This synchronization doesn't work well with
opening device on libary load (`dl_open`).

There is a deadlock between the main thread (loading the forge _C
library) and the `completion_queue_thread` in tt-metal.

The completion thread is trying to acquire lock on `dl_load_lock`,
which is already being held by the main thread (inside of the `dl_open`).
The reason why the completion thread is acquiring this lock, is to
register the destructor for the TLS variable `dispatch_cmd_and_event`.

Hence, to unblock further uplifts of tt-mlir (tt-metal) - removing
code triggering the opening of the device during library load. This
code is related to torch 2.0 integration, and is not actually used
currently.