[compile] introduce forge graph module
For the purpose of encapsulating multiple graphs for a given module,
`ForgeGraphModule` class is introduced.

The class for now simply stores pointers to graphs by the type of the
graph. Graph type can be one of the following: `Forward`, `Backward`,
`Loss`, `Optimizer`.

When lowering to mlir, we will iterate through all the graphs inside
the `ForgeGraphModule` and lower them as mlir functions.

This is a first change in a set of changes needed for separating
forward and backward graphs (for training) - Issue #100.

Closes #222