[training] splitting forward and backward graphs
A pass is added, which takes the forge graph and creates a
`ForgeGraphModule`, by splitting the forward and backward parts of the
graph.

Since we are currently executing forward and backward pass as separate
programs (functions) within a binary, splitting the graphs seems to be
the natural approach.

Forward graph is created by cloning the forward nodes of the original
graph, and adding output nodes for all the intermediaate results the
backward pass will need.

The backward graph is constructed from backward nodes of the original
graph and utilizing the intermediate results from the forward graph.
For each gradient an output node is created.

The `CompiledModel` is updated to have the state from both graphs.

In `lower_to_mlir()` removed the previous hacks - now we simply emit
mlir function for each graph in `ForgeGraphModule` and no special hacks
are needed for getting inputs/params/outputs of each graph.

This change enables MNIST Linear training to work.

Closes #100 #163