[training] add unit tests for split graph pass
Unit tests are added to test some of the invariants that need to hold
when extracting the forward and backward pass as separate graphs.

Added variants for training with recompute, however that part still
doesn't work (#366). However, the testing infra is in place, so once the
issue is fixed this test should hopefully work.

Changes to the `forge_graph_module.hpp` are only whitespace
(formatting).

Closes #363