[training] add torch gradient hooks (#421)
Currently, the loss and optimizer parts are ran on the CPU. In order to
run backward pass (on the device) we need to know the gradients coming
from the `loss.backward()`. Before this change, the gradients from the
backward part of the loss function were manually passed to the backward
pass by the user.

This change adds appropriate torch hooks which are triggered every time
a gradient is computed. The hook will then tie the calculated gradient
on the CPU to our runtime, so that it can be passed by the forge runtime
to the backward program when needed.

To discern between graph outputs which are used/defined by the user and
those which are used by us (to pass information between different
graphs/programs), output type enum is introduced. In this space, there
is still a lot of refactoring/redesigning to be done, once we get a
firmer grasp of what we want/need.

Closes #364