[train] support loss on device
Part of the changes for initial version of executing training e2e on
device. Changes to follow after this one will focus on the optimizer.

With this change, the loss module and the model can be compiled
separately and executed on the TT device (both forward and backward
passes). Example:

```python
tt_model = forge.compile(framework_model,
        sample_inputs=[inputs], training=True)

loss_fn = CrossEntropyLoss(name="cross_entropy_loss")
tt_loss = forge.compile(loss_fn, sample_inputs=loss_inputs,
        attach_to=tt_model, training=True)

 # Forward pass is executed as before
 #
 # The following will execute the whole backward pass from the loss
outputs down to the model backward pass.
tt_loss.backward()
```

Note: change to the API
- main compile function is modified to accept `training` parameter. To
indicate wheter to compile the module for training.
- also, loss module is removed as an argument to the compile function

For some reason, the gradient inputs in the previous stack were
represented as `InputNodeType::Loss`. I have added a new type of input
`Gradient`. The removal of the `Loss` input type is to be done as a
follow up change since there are some uses of it spread around in the
code base. Issue #829

To tie the gradients from the `loss.backward()` to the
`module.backward()` we need to "attach" the model to the loss module
when compiling the loss. This is done by passing the module to be
attached into the compile function (`attach_to` parameter).

Note: this doesn't work in general case, when there are multiple
gradients being passed between modules - because currently we don't have
a mechanism to know which gradient output to tie to which gradient
input.

Closes #177