[cleanup] removing global compiler configuration
This is a part of ongoing effort to remove any global state left over
from pybuda.

There should be no need of having global configuration of compiler. If
user wants to override some of the exposed settings, it should be done
locally, by creating a new (default) config and then passing it into the
compile function. All tests which need to modify the config should do
the same.

No feature should rely on reading from a "volatile global
configuration". Instead, each feature must either be designed to receive
the user-specified compiler configuration directly or defer
config-dependent decisions for later (until the compilation stage where
the appropriate configuration is available).

In majority of the cases i would assume that direct access to the user
passed config is easily available.

Closes #147