[build & ci] tt-forge-fe wheel
This change enables creating a `tt-forge-fe` wheel. Additionally, CI
`build-and-test` workflow is modified to have one job which creates a
python wheel (build) and then other jobs which are executing the tests
(using the previously generated wheel).

Logic for creating the wheel is contained in `setup.py`. It simply runs
cmake to build and install the project. The `setuptools` then picks up
and creates the wheel.

Installing `tt-mlir` via `cmake --install build --component SharedLib`
installs all needed dependencies for `tt-mlir` (including `tt-metal`).

`TT_METAL_HOME` environment variable is no longer needed, small piece of
logic in `forge/forge/__init__.py` tries to deduce it, relative to its
location. If it can't be deduced, an error message will be shown.

However, if the `TT_METAL_HOME` variable is set the logic will respect
it as long as it is a valid path, so overriding it in custom scenarios
is possible.

Closes #200