[runtime] tensors on device aka runtime stitching
This change implements a mechanism to keep certain tensors on device, so
that running programs in a loop (or chaining multiple programs) is more
efficient. We call this - runtime stitching.

Previously, we were sending each input to device each time we ran a
program - which is problematic, since this includes all of the weights,
which for example remain the same during multiple invocations of the
forward program (in inference) and don't need to be sent to the device
each time.

To keep the tensors on the device we only need to hold on to the
handles of the tensors provided by the device runtime (`tt-mlir`
runtime) so that the tensors don't get deallocated; and pass them on
when needed for executing a program. This change introduces several
structures for enabling us to that. Tensors which are kept on the device
are called `persistent`.

Summary of the new structures:
- `ModelState` which holds info for running the model (module), i.e.
  multiple programs (e.g. tensor pool, all program states)
- `ProgramState` - holds info needed for running a particular program
  (e.g. persistent input tensors).
- `TensorPool` simple container for holding named tensors; in the
  context of the `ModelState` it will contain all `persistent` tensors.
- `Tensor` which represents forge runtime tensor and holds both the
  handle to the runtime tensor (if exists) and the host storage handle -
  the torch tensor from which it was created. This tensor object is
  similar to the torch tensor, in the sense that the object contains
  (in this case `shared_ptr`) pointer to the actual tensor data. Thus,
  copying a tensor only creates a new handle to the underlaying data.

The rest of the change is minimal refactor of existing code to
incorporate new structures. Additionally, some of the training tests
were modified so that they remain functional; additional handling is
needed to ensure proper updating of host/device weight tensors, because
training tests combine executing some parts of the graphs/modules both
on the device and cpu.

There is no additional test coverage added and the correctness relies on
the existing e2e tests. More changes are yet to come with the goal of
creating a more performant and robust forge runtime. More fine grained
functional tests of the runtime component will be added.

Closes #1245, Closes #1329