[cleanup] minor refactoring and type hinting fixes
This cleanup is centered around `compiled_graph_state.py` and the code
it uses. I've taken slight detours to cleanup a bit more (e.g.
`compiler_cfg.py`, `module.py`) but generally tried to contain changes
as much as possible.

Removed `typing.py` as it was causing issues with circular dependencies
- now the types related to tensors are defined in `tensor.py`, and
  module related stuff is in `module.py`.

In `graph_lib/python_bindings.cpp`, some of the definitions are
reordered so that `mypy` can generate better stubs (for some reason the
order is important so that proper type hints are generated).