Revert 'reduce_avg' attribute from 'dim_arg' to 'dim' and update op_mapping (#1129)
### Problem description
The previous change in [PR
#1101](https://github.com/tenstorrent/tt-forge-fe/pull/1101) modified
the named attribute of `reduce_avg` from `dim` to `dim_arg` to align
with tt-mlir naming conventions. However, this change has been
identified as problematic and needs to be reverted.

### What's changed
This PR reverts the changes introduced in [PR
#1101](https://github.com/tenstorrent/tt-forge-fe/pull/1101), restoring
the named attribute of `reduce_avg` back to `dim`. Additionally, the
`op_mapping` has been updated from `dim` to `dim_arg` to ensure
compatibility with ttir.

### Checklist
- [x] New/Existing tests provide coverage for changes