[CI] reorganize workflows to use python wheel (#1160)
Before this change, we had `build-and-test` workflow which coupled
building and testing. Which means if we would want to split tests into
more groups or add more testing (both on `n150` and `n300`) each of
those jobs would waste time on building everything from scratch.

With this change we now have one `forge-build` job which creates both
`tt-tvm` and `tt-forge-fe` wheels, which will then be reused by all of
the testing jobs. Both `nightly` and `on PR` runs are udpated.

Additionally, we will run all tests on both `n150` and `n300` cards to
have better test coverage. In the past, we've had blind spots where some
of the tests were passing on `n150` and failing on `n300`.

In the future, the plan is to add workflows for `debug` build, test
coverage runs, etc.

Closes #200