[api] add API for exporting model as c++ source
Adding support for exporting compiled model as c++ source, via the
MLIR `emitc` pipeline.

To export the model, the user would call `export_to_cpp()` on the
already compiled model (`CompiledModel` instance). This means that
there are no changes in the compile flow.

The `CompiledModel` will hold the `ForgeGraphModule` generated during
the compile, so that it can execute the needed `tt-mlir` pipeline for
generating c++ code on demand. The other option would be to hold the
generated `ttir` text (during compile) in memory, so that the pipeline
can be ran from that point onwards. I've chosen the first option, since
it's simpler to fit in the existing flow, and I don't see a big
improvement in the other approach.

Existing code for running the MLIR compiler is refactored a bit to allow
generating different outputs (flatbuffer or cpp) without duplicating the
logic.

Simple test is added to confirm the C++ generation completes
successfully.

Closes #1340