[ci] remove build step from dockerfile (#1190)
There is some flakiness when building the docker image on the runners.
The job by itself needs around two hours to complete and it happens
occasionally that it somehow crashes the CI runner.

Removing the `tt-forge-fe` build step from the docker image. This is not
needed, since the docker image is afterwards used in CI, which will
verify that the build works properly.