[uplift] Switch to new dylib structure for ttmlir (#1154)
### Problem description
Making some changes to tt-mlir to avoid static linking. This entails
splitting TTMLIR.so into 2 parts, this is trivial PR to change so we
link both these parts (compiler + runtime) so instead of master so

### What's changed
Now using TTMLIRCompiler & TTMLIRRuntime libs instead of
TTMLIR in a cmake file.

Also, adding a temporary workaround in `setup.py`, because 
installation of `tt-metal` libs in `tt-mlir` is broken.