[ci] fix build docker image workflow
In cases where the `tt-mlir` commit hash is specified in workflows, we
don't respect it in the docker build job. So we end up with wrong docker
image in the rest of the workflow.

Fix: check if the `tt-mlir` commit hash is overriden and use it in the
docker build job.