Add job id to artifact name (#1421)
### Ticket
N/A

### Problem description
Github artifacts are not properly uploaded in workflow re-runs.
Artifacts are uploaded with same name and it is not possible to retrieve
them using GH api, it returns artifacts from first attempt.

### What's changed
Adding job_id to artifact name will make it unique across re-runs and
will allow us to download them.

### Checklist
- [x ] New/Existing tests provide coverage for changes