Update tvm commit hash retrieval for out-of-tree execution (#1393)
### Ticket
N/A

### Problem description
The tvm cache path retrieval assumed execution within the Forge tree,
causing an error when run out-of-tree as the tvm submodule could not be
found.

### What's changed
This change modifies the commit hash retrieval method to use the path
from `tvm.__file__`, allowing the cache to be saved correctly even in
out-of-tree scenarios.

### Checklist
- [ ] New/Existing tests provide coverage for changes