Write down the correct values for certain parameters in the LLAMA benchmark model. (#1461)
Some output parameters in the LLAMA benchmark model have inappropriate
default values and cannot be written to the database due to schema
constraints. In this PR, we update those values. The most important
parameters are the number of layers—which is irrelevant in this
phase—and image size, which cannot exist as an input in language models.