[signal handler] don't initialize TTSystem
During signal handling we want to close the opened device(s), so that
we leave the device(s) in a consistent state. However, if the `TTSystem`
singleton instance was not initialized before the signal handler is
called, the signal handler will end up initializing the system (i.e.
open/close the devices) unecessarily.

Fix this by checking if the system has been initialized before closing
the device(s); if not initialized - then we certainly don't have any
open devices at that point.

Also, move closing of the devices after the stacktrace has been printed;
this is done so that we'll have the stacktrace, even if something goes
wrong during the closing of the devices.