[runtime] add mechanism to enable program cache (#1525)
There are still some issues that need to be resolved before we enable
program cache on by default. For now, adding a mechanism for manually
enabling it in runtime.

Perf benchmark models are updated to enable program cache before
executing the benchmarks.

The `tt-forge-fe` APIs for configuring/managing the devices are still
not properly defined and are work in progress. This is a starting point
so that we can enable program cache in benchmarks and catch any
regressions coming from `tt-metal`/`tt-mlir`. Hence, adding these APIs
to the `experimental` submodule of `runtime` module.

Closes #1493