Add support for paddlepaddle and onnx framework in model analysis (#1666)
1. Added support for **paddlepaddle** and **onnx** framework in **model
analysis ops test generation pipeline**.
2. In the model analysis, remove the old `record_forge_property` fixture
and updated it with `forge_property_recorder` fixture.
3. Moved the forge_property_recorder pytest fixture function from
`tt-forge-fe/forge/test/conftest.py` path to `tt-forge-fe/fconftest.py`
because the model analysis markdown generation pipeline will generate
unique ops test in the `generated_modules/unique_ops/ `directory by
default but the pytest fixture is not visible when it is present inside
the `tt-forge-fe/forge/test/conftest.py`.
4. Created a pytest fixture(i.e `restore_package_versions`) which
ensures package versions remain consistent during tests execution and
added the fixture in yolov5 models tests. In the yolov5 model tests(i.e
`forge/test/models/pytorch/vision/yolo/test_yolo_v5.py`), the numpy
version is upgraded from 1.23.1 to 2.2.4 which causes below issues
```
AttributeError: _ARRAY_API not found

A module that was compiled using NumPy 1.x cannot be run in
NumPy 2.2.4 as it may crash. To support both 1.x and 2.x
versions of NumPy, modules must be compiled with NumPy 2.0.
Some module may need to rebuild instead e.g. with 'pybind11>=2.12'.

If you are a user of the module, the easiest solution will be to
downgrade to 'numpy<2' or try to upgrade the affected module.
We expect that some modules will need time to support NumPy 2.
```

5. Refactored the `ForgePropertyStore` and `ForgePropertyHandler` to
record the inputs and outputs tensor description inside the tags
property.
6. Removed the requires_grad=False setting in the Mistral and RCNN
models because it caused the forge module to treat model weights as
constants instead of parameters.
7. In the model analysis ops test generation, for advindex op, added
maximum int calculation which helps to generate indicies tensor with the
range of reference tensor.
8. Added support for dumping logs while extracting the unique ops
configuration for models in model analysis pipeline which helps for
debugging the test failures.