[verification] fix fallback condition (#1772)
With commit 22306940, custom kernels for calculating pcc/atol were
introduced. These kernels need to have inputs of same dtype and don't
support bool/uint8/int8 data types.

Fixing condition, under which we fallback to the previous calculation,
to include checks for (u)int8.

This fixes some of the model ops tests which are failing now.

Additionally, part of the change in `compare.py` is being reverted,
since I've originally misinterpreted the previous logic and created a
bug. Since the original logic doesn't make much sense, this will be
reworked in the future.