[onnx] don't create onnx session twice
When compiling ONNX models, we are creating the inference session twice.

Since this is not needed and the amount of memory required by the
session is not trivial - this change reuses the session which was
originally created.

Because the session and the `onnx_model` objects can be quite large, I
am calling the `del` operator explicitly, to minimize their lifetime.

Additionally, removing the unused code related to `dequantize`.

Closes #1800