Enable lower data format user override. (#1805)
### Ticket
Closes #1804 

### Problem description
Optimizer needs a way to use lower precision data formats. Since TVM
requires traced model to be in float32 (due to its numpy base),
generated initial graph is in float32. To transfer this to bfloat16 (or
some other lower data format) two things needed to be done:
1. cast all input tensors (Inputs, weights, constants) to lower data
format.
2. convert output data format in whole graph

We used existing config `default_df_override`. By setting this to some
data format, compiler will cast the input tensors and operations to that
lower data format. Therefore enabling optimizer to have smaller memory
footprint and enable higher batch size.

### What's changed
- Set user-defined data format override in post-initial graph pass.
Insert cast nodes on all input edges to ensure that constants, buffers,
and parameters are in the lowered dtype.
- Move TVM module generation inside the block where the traced model is
in float32, since that is the format supported by TVM (due to its NumPy
base).
- Extend ConvertEmulatedDtypes to convert model buffers (e.g., BatchNorm
runing mean and variance) to the desired dtype. Buffers are not
considered parameters but are inputs to the computational graph and must
be aligned in data format with the rest of the model.

- Added pytorch tests for casting model to bfloat16.

### 
What is not done:

Casting model to bfp8_b. There are some concerns regarding bfp8_b
implementation in forge (mainly because torch doesn't support float8 and
block formats and we rely on torch for verification and model tracing
before TVM). Also currently there is no support for lowering bfp8_b to
mlir.