[compile] optimize memory usage for cpp->py calls
In majority of graph passes we are iterating over the whole graph and
calling python functions (from c++) to check various things (is the op
element-wise, is it a tm, etc.). Each of these calls was copying the
passed `OpType` struct, and allocating the copy on the heap (by pybind).

Huge number of these allocations is causing extensive memory usage and
memory fragmentation issues.

We can call python functions while passing the `reference_wrapper<T>`
via `std::ref()` to enable pybind to avoid copying the objects.

Changes:
- went through all places where we are passing `OpType` and replaced it
  `std::ref()` instead
- refactored the commonly used `is_eltwise()` to be a method on `OpNode`
- removed "forge.op.eval.lforge" usage, since the `lforge` module
  doesn't exist (artifact from the previous stack, hasn't been
  completely removed, yet)

For more details on reproducing the issue, see #1864.

Closes #1864