[mlir] enable using tt-mlir optimizer
To be able to run the optimizer, it is required to build `tt-mlir`
with certain flags - hence, passing these flags to the build.

The knobs to enable optimizer, memory layout analysis are added.

Also, there is a way to provide a custom string to be passed to MLIR -
for use in testing or for configs not properly exposed.

Made all config fields `std::optional`, so that user can specify only
one config flag, and the rest will not be specified in the pipeline
options string, i.e. they'll have their default values defined in mlir.

Enabled the optimizer and memory layout analysis on MNIST benchmark -
however, due to an issue we need to disable the device program cache.

Also optimizer without memory layout analysis is enabled for Resnet
benchmark.