[mlir] set argument type when lowering to mlir (#1870)
The attribute `tt.argument_type` is used to distinguish between model
parameters (weights), constants or regular inputs (which are expected to
change on each forward call).

This attribute is used in consteval logic in `tt-mlir` to hoist parts of
the graph which can be evaluated only once in runtime, cached on device
and reused afterwards (in subsequent program invocations).

Closes #1867