[mlir] basic support for configuring MLIR pipeline (#1873)
Adding structure `MLIRConfig` to enable setting some of the pipeline
options in `tt-mlir`. This structure is part of the `CompilerConfig` and
is set to `None` by default.

For now, the only config we expose is the one controlling consteval in
`tt-mlir`. Settings for enabling optimizer will follow.

Also, marking tests in `test_api.py` to be ran on `push`.

Closes #1869