[deps] bump version of fmt to 11.1.4 (#1884)
Recently, `tt-metal` bumped their version of `fmt` library. This causes
weird issues when running `tt-forge-fe` debug build. When opening the
device, a runtime exception `invalid format specifier` is thrown. The
issue is not observed in release builds.

I haven't root-caused the issue fully, but it looks like the code from
our `fmt` includes (ver. 11.0.2) in `third_party/fmt` is clashing with
their `fmt` includes on their newer version.

Bumping the version of `fmt` which we are using, which seem to resolve
the issues.