[mlir] construct hierarchical locations (#1908)
The current logic for creating locations, forms them in the following
way:

`NameLoc(graph_name, child_loc)` - where graph_name is in most cases
`"forward"` and `child_loc` contains the actual location:
`module/layer1/layer2/node_name`.

There are two problems with this:
1. `NameLoc` contains only value `"forward"` as its value, so we cannot
discern between nodes by looking at it (we would need to get the child
location instead)
2. even if the direct location would contain the layer information, the
string is far too large

Since MLIR supports nesting of locations, logic for creating the
locations is modified so that it parses the location string and creates
nested instances of `NameLoc`. In this way, we have the `NameLoc`
attached to a particular op containing its name. If we wish to see more
details about where this particular op belongs, we can go through the
chain of locations and reconstruct everything.

Example (location string -> `NameLoc`):

`"module_name/layer/op_name"` ->
`NameLoc("op_name", NameLoc("layer", NameLoc("module_name")))`

Also, modify the way we dump the module for `reportify` so that the
locations are included in the dump.