Removed forge_property_handler form cpp code.
ExecutionDepth (tests info) was used as a compiler substage execution information and it was implemented in cpp code.
It is moved to python.

Since we don't need to separate generating TTIR and MLIR compiler pass, both are saved as a single
FAILED_TTMLIR_COMPILATION execution depth.

Also, getting flatbuffer logic is simplified and it is moved to python code.

This is second part of forge property handler simplification.