[test] try to release memory after each test (#2046)
Run garbage collection and call `malloc_trim()` after each test.

`malloc_trim(0)` forces releasing of any unused memory to the OS.

The growth of `pytest` process memory usage as each test gets executed
is much lower now, but the memory usage still grows in time. There
doesn't seem to be any significant memory leaks on our side, so it looks
that some state which is persisted across tests is contributing to the
memory usage growth. This still needs to be investigated.

Memory usage logs are also updated with new columns so that we can track
the memory usage after gc and trimming.

Closes #2044, Closes #2045