Simplified forge property recorder/handler usage.
Forge property recorder/handler (FPH) was passed as parameter throughout all code base
in order to collect various forge properties.

With this change, FPH is initialized as a context variable at the beggining of every test
(with pytest fixture) and can be used with global recording functions that are introduced.
Every function checks whether FPH is initialized, and only records properties if it is.
This way, we have readable code with recording enabled only for tests.

This is the third part of forge property handler simplification.