Fix ssd_vgg16 Unsupported c++ op issue (#2107)
### Ticket
(https://github.com/tenstorrent/tt-forge-fe/issues/2033)

### Problem description
The ` RuntimeError: Couldn't load custom C++ ops.` encountered during
the execution of the framework model is due to an unsupported C++
operation. Specifically, the issue arises from the presence of the nms
(Non-Maximum Suppression) operation in the model's forward pass, which
is implemented in C++ and is currently not supported in this context.

### What's changed
This PR includes the changes to make the postprocessing outside the
model which resolves the issue.

Log:


[model_compilation_ssd_vvg16.log](https://github.com/user-attachments/files/20448786/model_compilation_ssd_vvg16.log)