Fix CI import issues in paddlenlp (#2204)
The **aistudio-sdk** package version `0.3.0`, released this morning,
removed the `download` function from hub.py, which is still used by the
**paddlenlp** package. This caused an` ImportError: cannot import name
'download' from 'aistudio_sdk.hub'` in CI pipeline. During the
installation of **paddlenlp** version` 2.8.1`, the latest aistudio-sdk
was pulled due to the `aistudio-sdk>=0.1.3` requirement specified in its
requirements.txt. To resolve the issue, we reverted aistudio-sdk from
`0.3.0` back to version `0.2.6`.