Enable lower dataformat support in models ops gen (#2218)
Currently we are running convolution based models in **lower
dataformat(eg: float16_b)**. For running the models in lower dataformat,
we will convert the **model** and **inputs** to **lower dataformat(eg:
bfloa16)** and then pass it to compile function but the **tvm graph** is
constructed on **float32** because the `ConvertEmulatedDtypes` convert
the **weights**, **buffers** and **inputs** from **bfloat16** to
**float32**. In the `post initial graph pass`, we will convert the
**parameter** and **op node** dataformat to **lower dataformat(eg:
float16_b)** and add **cast node** after the **constant** node if the
dataformat is not **lower dataformat(eg: float16_b)** and we are already
passing activation and parameter as bfloat16 so there is no need to
lower.

Since the lower dataformat conversation is happening only in post intial
graph pass there will be no change in **ForgeModule generated
with/without lower dataformat** because it is constructed on the **TVM
graph.**

Similarly there will be no change in the **models ops generation**
because the **unique ops config** is extracted from the **TVM graph**.

The PR will enable **lower dataformat support** in the **models ops
tests generation pipeline** to match the same config as convolution
models.

It will convert all the **parameters**, **constant** and **activation**
from **floating dtype(eg: float32)** to **lower dataformat(eg:
float16_b/bfloat16**) and will also save the lower dataformat as
metadata for setting the compiler config default_df_override so that
`configure_output_data_formats` function in `post initial graph pass`
will assert if the constant and activation are not in lower dataformat
and convert the parameter and op node to lower dataformat if is missed.